backend/regen-thumbnails/requirements.txt
Pillow>=10.0.0
boto3>=1.26.0
psycopg2-binary>=2.9.0
